Relocate profitbricks util module
