+++ .github/BOTMETA.yml
@@ -449,6 +449,9 @@ files:
       - cloud
   $module_utils/acme.py:
     maintainers: resmo felixfontein
+  $module_utils/cloud/profitbricks/profitbricks.py:
+    maintainers: $team_profitbricks
+    labels: profitbricks
   $module_utils/cloudstack.py:
     maintainers: $team_cloudstack
     labels: cloudstack
@@ -1076,6 +1079,7 @@ macros:
   team_onyx: samerd
   team_openstack: emonty omgjlk juliakreger rcarrillocruz shrews thingee dagnello
   team_openswitch: Qalthos gdpak
+  team_profitbricks: edevenport nurfet-becirevic
   team_rabbitmq: chrishoffman manuel-sousa hyperized
   team_rhn: alikins barnabycourt flossware vritant
   team_scaleway: sieben hekonsek Spredzy
